DB path now in appsetting.json and add test db path update DB connection with new paths start DB incerts method

+++ Models/AnimeModel.cs
@@ -4,7 +4,7 @@
     {
         public required string OriginalName { get; set; }
         public required bool Country { get; set; } // true = Japan | false = China
-        public required int Rank { get; set; } // 1-10 and 11 = S
+        public required byte Rank { get; set; } // 1-10 and 11 = S
         public required int Episodes { get; set; } // if only 1 is a Anime Movie
 
         public required ItemRawModel Item { get; set; }
